Moved kernel into tool file to make it standalone
